完成添加修改删除采购付款接口

+++ psi-java/psi-domain/src/main/java/com/zeroone/star/project/dto/paymentmanagement/purchasepayment/PaymentDeleteDTO.java
@@ -2,6 +2,8 @@ package com.zeroone.star.project.dto.paymentmanagement.purchasepayment;
 
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
+import io.swagger.models.auth.In;
+import lombok.AllArgsConstructor;
 import lombok.Data;
 
 /**
@@ -12,5 +14,5 @@ import lombok.Data;
 @ApiModel("删除采购付款的数据传输对象")
 public class PaymentDeleteDTO {
     @ApiModelProperty(value = "单据id",example = "1625823227778736130")
-    private int id;
+    private Long id;
 }
